requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sentence-transformers>=2.7.0
faiss-cpu==1.7.4
numpy>=1.24.3
openai>=1.12.0
torch>=2.2.0
python-dotenv==1.0.0
transformers>=4.35.0
huggingface_hub>=0.20.0

